pyproject.toml
[tool.poetry]
name = "Soundscribe WebUI"
version = "0.1.0"
description = "A minimal web application with Soundscribe installed"
authors = ["Dr. Robert Turnbull <robert.turnbull@unimelb.edu.au>", "James Quang <james.quang@unimelb.edu.au>"]
license = "Apache Software License 2.0"
readme = "README.rst"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10,<4.0"
django-watson = ">=1.5.5"
django-stronghold = ">=0.4.0"
rcssmin = ">=1.1.0"
argon2-cffi = ">=21.1.0"
whitenoise = ">=5.3.0"
redis = ">=3.5.3"
hiredis = ">=2.0.0"
celery = ">=5.1.2"
setuptools = ">=68.0.0"
django-celery-beat = ">=2.2.1"
django-timezone-field = ">=4.2.1"
flower = ">=1.0.0"
django-environ = ">=0.8.1"
django-model-utils = ">=4.2.0"
django-allauth = ">=0.63.0"
django-crispy-forms = ">=1.13.0"
django-compressor = ">=2.4.1"
django-redis = ">=5.0.0"
djangorestframework = ">=3.12.4"
django-cors-headers = ">=3.10.0"
django-extensions = ">=3.1.3"
django-filer = ">=2.0.2"
django-cms = ">=3.9.0,<=3.11.6"
djangocms-admin-style = ">=2.0.2"
djangocms-attributes-field = ">=2.0.0"
djangocms-file = ">=3.0.0"
djangocms-googlemap = ">=2.0.0"
djangocms-link = "4.0.0"
djangocms-picture = ">=3.0.0"
djangocms-snippet = ">=3.0.0"
djangocms-style = ">=3.0.0"
djangocms-text-ckeditor = ">=4.0.0"
djangocms-video = ">=3.0.0"
djangocms-bootstrap4 = ">=2.0.0"
djangocms-page-meta = ">=1.0.1"
django-storages = {extras = ["boto3"], version = ">=1.12.2"}
gunicorn = ">=20.1.0"
django-anymail = ">=8.4"
psycopg2-binary = ">=2.9.1"
djangocms-unimelb = {git = "https://gitlab.unimelb.edu.au/mdap-public/djangocms-unimelb.git"}
django-mptt = ">=0.14.0"
django-guardian = ">=2.4.0"
numpy = ">=1.26.4"
django-mdeditor = ">=0.1.20"
markdown = ">=3.5.2"
crispy-bootstrap4 = "^2024.1"
ffmpeg-python = "^0.2.0"
chardet = "^5.2.0"
soundscribe = {git = "git@github.com:unimelbmdap/soundscribe.git", branch = "main"}

[tool.poetry.group.dev.dependencies]
Werkzeug = ">=2.0.2"
ipdb = ">=0.13.9"
python-dotenv = ">=0.19.1"
notebook = ">=6.4.5"
mypy = ">=0.910"
django-stubs = ">=1.9.0"
pytest = ">=6.2.5"
pytest-sugar = ">=0.9.4"
Sphinx = ">=4.2.0"
sphinx-autobuild = ">=2021.3.14"
sphinx-rtd-theme = ">=0.5.1"
nbsphinx = ">=0.8.7"
flake8 = ">=3.8.4"
flake8-isort = ">=4.0.0"
coverage = ">=5.4"
black = ">=21.9b0"
pylint-django = ">=2.4.4"
pylint-celery = ">=0.3"
pre-commit = ">=2.15.0"
factory-boy = ">=3.2.0"
django-debug-toolbar = ">=3.2.2"
django-coverage-plugin = ">=2.0.1"
pytest-django = ">=4.4.0"

[build-system]
# Should be the same as `$POETRY_VERSION`:
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"
